Update README.md
DriverPass Project
The project aims to assist many people failing in their Department of Motor Vehicle exams and enabling people to study online without attending physical classes. Its main objective is to create a system that will allow clients to register, change passwords automatically and have full access to learning materials from the Department of Motor Vehicles. The client is Mr. Liam, the owner of Driver Pass Company that teaches basic driving rules in the Department of Motor Vehicles. The owner wants to create an application that will contain all the content he teaches physically in class. The application is designed to be used by learners who prefer to study online than physically. All users of the application can make appointments and share their experiences about the online services offered by the company.
Ensuring safety and adaptability when accessing the system has been performed well. To ensure security, the user must have a username and password when accessing the system. After an employee leaves the company, their access will be restricted. Additionally, when there are many failed login attempts, the system will automatically lock the user's account. Driver Pass also ensured that if users cannot remember their pin, they must automatically enter the system and reset it without needing a secretary's assistance. The adaptability of the systems allows modifications to be made whenever necessary without having to change the codes. Easy modifications aimed to ensure that the process was simple to perform because there were no system analysts. The system is made so that the updates are according to how the business develops, and when they need to make updates, it is simple since the system is readily designed for any updates.
An area that needs revision is the part of staff required to conduct this project successfully. Based on the intensity of the project, an improvement required is the employment of several people to enable the completion of all activities in the program. System analysts are also crucial for they offer technical expertise in managing and maintaining the systems. The system experts can modify the system to allow the users to change their username and also be able to retrieve it easily. Additionally, with the employment of several experts, users can consult in areas they need clarification while using the application.
In designing, it is essential to meet the user's needs. The user need was an application that could perform the work he does in a physical class. Based on this need, I preferred the cloud, which is fast and capable of running on the web. Mr. Liam wants to focus on his business; therefore, the cloud can offer security for his important data, unlike the application that would make him have a place for storing information manually. Additionally, it guides in the making of an affirmative user experience that suits a unique target. Understanding users is a fundamental requirement in running any system successfully.
My approach in designing software is the Function Oriented Approach which is a tactic to software design in which a design is disintegrated into a set of interacting units with a clearly defined function. The approach begins with an advanced description of the functions of the software. Furthermore, it breaks down each section of the report, describing its usefulness in greater depth. The best technique to use in the future is the Data Flow Diagram (DFD which depicts how important information flows across a system. It shows the storage sites, data inputs, outputs, and paths between each destination using predetermined symbols like rectangles, circles, arrows, and small text labels.